autocommit/pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["stong <498327580@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.9"
python-dotenv = "^1.0.0"
gitpython = "^3.1.32"
unidiff = "^0.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
